Refactored new-user form
