+++ src/app/dashboard/users/new-user.js
@@ -1,9 +1,7 @@
 // @flow
 import React, { Component } from 'react';
-import { Row, Col, Button } from 'react-bootstrap';
 import { setValue, getValue } from 'neoform-plain-object-helpers';
 import UserForm from './components/user-form.js';
-import UserForm2 from './components/user-form2.js';
 
 export default class NewUser extends Component<any> {
 
@@ -38,7 +36,7 @@ export default class NewUser extends Component<any> {
 
     return (
       <div>
-        <UserForm2
+        <UserForm
             data={this.state}
             getValue={getValue}
             onChange={this.onChange}
